requirements.txt
Django==4.2
django-redis==5.4.0
psycopg2-binary==2.9.9
djangorestframework==3.15.1
django-filter==24.2
django_model_utils==4.5.1
drf-dynamic-read==0.1.1

